Fixed dispute info
