+++ src/pages/covers/[coverId]/products/[productId]/index.jsx
@@ -1,13 +1,13 @@
 import Head from 'next/head'
 import { CoverOptionsPage } from '@/src/modules/cover/CoverOptionsPage'
-import { isV2BasketCoverEnabled } from '@/src/config/environment'
+import { isDiversifiedCoversEnabled } from '@/src/config/environment'
 import { ComingSoon } from '@/common/ComingSoon'
 import { useRouter } from 'next/router'
 import { useCoverOrProductData } from '@/src/hooks/useCoverOrProductData'
 import { safeFormatBytes32String } from '@/utils/formatter/bytes32String'
 import { isValidProduct } from '@/src/helpers/cover'
 
-const disabled = !isV2BasketCoverEnabled()
+const disabled = !isDiversifiedCoversEnabled()
 
 export default function Options () {
   const router = useRouter()
