+++ src/pages/reports/[coverId]/products/[productId]/incidents/[timestamp]/dispute.jsx
@@ -1,18 +1,9 @@
 import Head from 'next/head'
 import { useRouter } from 'next/router'
-import { useFetchReport } from '@/src/hooks/useFetchReport'
-import { NewDisputeReportForm } from '@/src/modules/reporting/NewDisputeReportForm'
-import { ReportingHero } from '@/src/modules/reporting/ReportingHero'
-import { Container } from '@/common/Container/Container'
-import { Alert } from '@/common/Alert/Alert'
-import DateLib from '@/lib/date/DateLib'
-import { isGreater } from '@/utils/bn'
 import { ComingSoon } from '@/common/ComingSoon'
 import { isFeatureEnabled } from '@/src/config/environment'
-import { Trans } from '@lingui/macro'
-import { CoverStatsProvider } from '@/common/Cover/CoverStatsContext'
 import { safeFormatBytes32String } from '@/utils/formatter/bytes32String'
-import { useCoverOrProductData } from '@/src/hooks/useCoverOrProductData'
+import { NewDisputeReportFormContainer } from '@/modules/reporting/NewDisputeReportFormContainer'
 
 const disabled = !isFeatureEnabled('reporting')
 
@@ -22,83 +13,21 @@ export default function DisputeFormPage () {
   const coverKey = safeFormatBytes32String(coverId)
   const productKey = safeFormatBytes32String(productId || '')
 
-  const coverInfo = useCoverOrProductData({
-    coverKey: coverKey,
-    productKey: productKey
-  })
-  const { data, loading } = useFetchReport({
-    coverKey: coverKey,
-    productKey: productKey,
-    incidentDate: timestamp
-  })
-
-  if (!coverInfo) {
-    return <Trans>loading...</Trans>
-  }
-
   if (disabled) {
     return <ComingSoon />
   }
 
-  const now = DateLib.unix()
-  const reportingEnded = data?.incidentReport
-    ? isGreater(now, data.incidentReport.resolutionTimestamp)
-    : false
-
-  const canDispute =
-    !reportingEnded && data?.incidentReport?.totalRefutedCount === '0'
-
   return (
-    <CoverStatsProvider coverKey={coverKey} productKey={productKey}>
-      <main>
-        <Head>
-          <title>Neptune Mutual Covers</title>
-          <meta
-            name='description'
-            content='Get guaranteed payouts from our parametric cover model. Resolve incidents faster without the need for claims assessment.'
-          />
-        </Head>
-
-        {/* hero */}
-        <ReportingHero
-          coverKey={coverKey}
-          productKey={productKey}
-          coverInfo={coverInfo}
-          reportStatus={{
-            resolved: data.incidentReport?.resolved,
-            dispute: true
-          }}
+    <main>
+      <Head>
+        <title>Neptune Mutual Covers</title>
+        <meta
+          name='description'
+          content='Get guaranteed payouts from our parametric cover model. Resolve incidents faster without the need for claims assessment.'
         />
-        <hr className='border-t border-t-B0C4DB' />
+      </Head>
 
-        {loading && (
-          <p className='text-center'>
-            <Trans>loading...</Trans>
-          </p>
-        )}
-
-        {!loading && !data.incidentReport && (
-          <p className='text-center'>
-            <Trans>No data found</Trans>
-          </p>
-        )}
-
-        {data.incidentReport && (
-          <div>
-            {canDispute
-              ? (
-                <NewDisputeReportForm incidentReport={data.incidentReport} />
-                )
-              : (
-                <Container className='py-16'>
-                  <Alert>
-                    <Trans>Not applicable for disputing</Trans>
-                  </Alert>
-                </Container>
-                )}
-          </div>
-        )}
-      </main>
-    </CoverStatsProvider>
+      <NewDisputeReportFormContainer coverKey={coverKey} productKey={productKey} timestamp={timestamp} />
+    </main>
   )
 }
