+++ src/styles/globals.css
@@ -2,19 +2,17 @@
 @tailwind components;
 @tailwind utilities;
 
-@layer base {
-  [type="button"] {
-    -webkit-appearance: none;
-  }
+:root {
+  accent-color: #4e7dd9;
+}
 
-  [type="radio"]:focus,
-  [type="radio"]:checked {
-    box-shadow: none;
-    @apply !bg-4e7dd9 outline outline-1 outline-B0C4DB outline-offset-0 bg-none border-4 !border-white;
+@layer base {
+  [type='button'] {
+    -webkit-appearance: none;
   }
 }
 
-*[tabindex="-1"] {
+*[tabindex='-1'] {
   outline: none;
 }
 
@@ -49,7 +47,7 @@ input:checked + div svg {
     rgba(233, 233, 233, 0.8) 100%
   );
   animation: shimmer 1s ease-out infinite;
-  content: "";
+  content: '';
 }
 
 @keyframes shimmer {
